Terminate task after posting semaphore
If switch the task state to "TERMINATED" before
posting semaphore, scheduler may dequeue the
task from active queue, therefore make the
task unable to be scheduled even that it
not post semaphore yet. Other task waiting on
the task will not be able to finish.

To fix, I move the statement which set
the task to "TERMINATED" after posting semaphore
statement.